pyproject.toml
[project]
name = "sketch-vs-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Simplifying-Document-Processing-with-UiPath-and-ChatGPT",email = "orooba80@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "openai (>=1.97.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sketch-vs-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

packages = [
    { include = "app" }
]